Trading View/tv-tradingviewwithpostgreDB_nifty50/requirement.txt
websocket_client~=0.57.0
SQLAlchemy~=1.3.20
configparser~=5.0.1
coloredlogs~=14.0